refactor class to androidx and rename package
