+++ app/src/main/java/io/github/rahsyarigami/infonunukan/data/model/ItemKontak.java
@@ -1,4 +1,4 @@
-package io.github.rahsyarigami.infonunukan.model;
+package io.github.rahsyarigami.infonunukan.data.model;
 
 public class ItemKontak {
     private String itemListContact;
