+++ app/src/main/java/io/github/rahsyarigami/infonunukan/data/model/ItemKuliner.java
@@ -1,4 +1,4 @@
-package io.github.rahsyarigami.infonunukan.model;
+package io.github.rahsyarigami.infonunukan.data.model;
 
 public class ItemKuliner {
