+++ app/src/main/java/io/github/rahsyarigami/infonunukan/ui/adapter/KontakPentingAdapter.java
@@ -1,4 +1,4 @@
-package io.github.rahsyarigami.infonunukan.view.adapter;
+package io.github.rahsyarigami.infonunukan.ui.adapter;
 
 import android.content.ClipData;
 import android.content.ClipboardManager;
@@ -6,9 +6,7 @@ import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.RecyclerView;
+
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -16,16 +14,20 @@ import android.widget.ImageButton;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
+import androidx.recyclerview.widget.RecyclerView;
+
 import java.util.List;
 
 import io.github.rahsyarigami.infonunukan.R;
-import io.github.rahsyarigami.infonunukan.model.ItemKontak;
+import io.github.rahsyarigami.infonunukan.data.model.ItemKontak;
 
 
 public class KontakPentingAdapter extends RecyclerView.Adapter<KontakPentingAdapter.KontakViewHolder> {
 
-    private Context mContext;
-    private List<ItemKontak> mItemKontak;
+    private final Context mContext;
+    private final List<ItemKontak> mItemKontak;
     // private static final int PERMISSION_REQUEST_CODE = 200;
