+++ app/src/main/java/io/github/rahsyarigami/infonunukan/ui/detail/KontakPenting.java
@@ -1,16 +1,18 @@
-package io.github.rahsyarigami.infonunukan.view.activity;
+package io.github.rahsyarigami.infonunukan.ui.detail;
 
 import android.annotation.SuppressLint;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.SearchView;
+
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.RelativeLayout;
+
+import androidx.appcompat.widget.SearchView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.facebook.shimmer.ShimmerFrameLayout;
 
@@ -19,9 +21,9 @@ import java.util.List;
 import java.util.Objects;
 
 import io.github.rahsyarigami.infonunukan.R;
-import io.github.rahsyarigami.infonunukan.model.ItemKontak;
-import io.github.rahsyarigami.infonunukan.util.BaseActivity;
-import io.github.rahsyarigami.infonunukan.view.adapter.KontakPentingAdapter;
+import io.github.rahsyarigami.infonunukan.data.model.ItemKontak;
+import io.github.rahsyarigami.infonunukan.ui.base.BaseActivity;
+import io.github.rahsyarigami.infonunukan.ui.adapter.KontakPentingAdapter;
 
 public class KontakPenting extends BaseActivity implements SearchView.OnQueryTextListener {
 
@@ -37,7 +39,7 @@ public class KontakPenting extends BaseActivity implements SearchView.OnQueryTex
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        crashlyticsView();
+//        crashlyticsView();
         setContentView(R.layout.activity_kontak_penting);
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
