+++ app/src/main/java/io/github/rahsyarigami/infonunukan/ui/detail/TentangNunukan.java
@@ -1,19 +1,20 @@
-package io.github.rahsyarigami.infonunukan.view.activity;
+package io.github.rahsyarigami.infonunukan.ui.detail;
 
 import android.annotation.SuppressLint;
 import android.os.Build;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
 import io.github.rahsyarigami.infonunukan.R;
-import io.github.rahsyarigami.infonunukan.model.ItemTentang;
-import io.github.rahsyarigami.infonunukan.util.BaseActivity;
-import io.github.rahsyarigami.infonunukan.view.adapter.TentangNunukanAdapter;
+import io.github.rahsyarigami.infonunukan.data.model.ItemTentang;
+import io.github.rahsyarigami.infonunukan.ui.base.BaseActivity;
+import io.github.rahsyarigami.infonunukan.ui.adapter.TentangNunukanAdapter;
 
 public class TentangNunukan extends BaseActivity {
 
@@ -25,7 +26,7 @@ public class TentangNunukan extends BaseActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        crashlyticsView();
+//        crashlyticsView();
         setContentView(R.layout.activity_tentang_nunukan);
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
