+++ app/src/main/java/io/github/rahsyarigami/infonunukan/util/decoration/CustomDividerItemDecoration.java
@@ -1,4 +1,4 @@
-package io.github.rahsyarigami.infonunukan.decoration;
+package io.github.rahsyarigami.infonunukan.util.decoration;
 
 import android.content.Context;
 import android.content.res.Resources;
@@ -6,11 +6,13 @@ import android.content.res.TypedArray;
 import android.graphics.Canvas;
 import android.graphics.Rect;
 import android.graphics.drawable.Drawable;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+
 import android.util.TypedValue;
 import android.view.View;
+
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 
 public class CustomDividerItemDecoration extends RecyclerView.ItemDecoration {
@@ -21,10 +23,10 @@ public class CustomDividerItemDecoration extends RecyclerView.ItemDecoration {
     private static final int HORIZONTAL_LIST = LinearLayoutManager.HORIZONTAL;
     private static final  int VERTICAL_LIST = LinearLayoutManager.VERTICAL;
 
-    private Drawable divider;
+    private final Drawable divider;
     private int mOrientation;
-    private Context context;
-    private int margin;
+    private final Context context;
+    private final int margin;
 
     public CustomDividerItemDecoration(Context context, int orientation, int margin){
         this.context = context;
