+++ app/src/main/res/layout-v21/activity_main.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorLightGrey"
-    tools:context=".MainMenu">
+    tools:context=".ui.main.MainMenu">
 
 
     <com.google.android.material.appbar.AppBarLayout
