+++ app/src/main/res/layout/activity_info_wisata.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/colorLightGrey"
     android:orientation="vertical"
-    tools:context=".view.activity.InfoWisata">
+    tools:context=".ui.detail.InfoWisata">
 
     <androidx.cardview.widget.CardView
         android:id="@+id/cardView"
