+++ app/src/main/res/layout/activity_kuliner_do.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:background="@color/colorLightGrey"
     android:orientation="horizontal"
-    tools:context=".view.activity.KulinerDO">
+    tools:context=".ui.detail.KulinerDO">
 
     <com.facebook.shimmer.ShimmerFrameLayout
         android:id="@+id/shimmer_kuliner"
