+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorLightGrey"
-    tools:context=".MainMenu">
+    tools:context=".ui.main.MainMenu">
 
     <com.google.android.material.appbar.AppBarLayout
         android:layout_width="match_parent"
