+++ app/src/main/res/layout/layout_gridview.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".MainMenu"
+    tools:context=".ui.main.MainMenu"
     tools:showIn="@layout/activity_main">
 
     <com.google.android.material.card.MaterialCardView
